+++ src/protocol/functions/Force_Single_Coil.js
@@ -41,7 +41,7 @@ var ForceSingleCoil = function (pdu){
     }
     else {
         let values = new Map();
-        this.coils.WriteData(value, targetCoil);
+        this.coils.SetValue(value, targetCoil);
         values.set(targetCoil, (value > 0));
         //creando una copia del request
         respPDU = pdu;