+++ src/protocol/functions/Mask_Holding_Register.js
@@ -30,14 +30,13 @@ var MaskHoldingRegister = function (pdu){
         let OR_MASK = pdu.modbus_data.readUInt16BE(4);
 
         //calculating value
-        let mask_value = (this.holdingRegisters.GetRegister(targetRegister).readUInt16LE() & AND_MASK) | (OR_MASK & ~AND_MASK );
-        let val = Buffer.alloc(2);
-        val.writeUInt16LE(mask_value);        
-        this.holdingRegisters.SetRegister(val, targetRegister);
+        let mask_value = (this.holdingRegisters.GetValue(targetRegister) & AND_MASK) | (OR_MASK & ~AND_MASK );
+             
+        this.holdingRegisters.SetValue(mask_value, targetRegister);
 
         //Devolviendo un eco de la pdu.
         respPDU = pdu
-        values.set(targetRegister, this.holdingRegisters.GetRegister(targetRegister).readUInt16LE());
+        values.set(targetRegister, this.holdingRegisters.GetValue(targetRegister));
         this.emit('values', '4x', values);
         return respPDU;