+++ src/protocol/functions/Read_Coil_Status.js
@@ -29,7 +29,7 @@ var ReadCoilStatus = function (pdu) {
         //cantidad de coils a leer
         var numberOfCoils =   pdu.modbus_data.readUInt16BE(2);
 
-        let masks = [0x01, 0x02, 0x04, 0x08, 0x010, 0x20, 0x40, 0x80];
+        //let masks = [0x01, 0x02, 0x04, 0x08, 0x010, 0x20, 0x40, 0x80];
 
         //Calculando cantidad de bytes de la respuesta 12%8=1
         //ejemplo 12 coils necesitan 2 bytes
@@ -40,14 +40,16 @@ var ReadCoilStatus = function (pdu) {
         respPDU.modbus_data[0]=byte_count;
 
         //buffer temporal con tamano suficiente para copiar el segmento del registro con las coils solicitadas
-        var coil_segment =  Buffer.alloc(byte_count);
-
+        //var coil_segment =  Buffer.alloc(byte_count);
+        var coil_segment = this.coils.EncodeRegister(initCoil, numberOfCoils);
+        /*
         for(var i = 0; i < numberOfCoils; i++){
           if(this.coils.ReadData(initCoil + i)){
             coil_segment[Math.floor(i/8)] = coil_segment[Math.floor(i/8)] | masks[i%8];
           }
           else coil_segment[Math.floor(i/8)] = coil_segment[Math.floor(i/8)] & (~masks[i%8]);
         }
+        */
 
         //copiando las cois al campo de data de la PDU
         coil_segment.copy(respPDU.modbus_data,1);